+++ config/initializers/external_apis/datacite.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+Rails.configuration.x.datacite.active = true
+Rails.configuration.x.datacite.api_base_url = 'https://api.datacite.org'